+++ noodles-bam/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * bam: Raise minimum supported Rust version (MSRV) to 1.89.0.
+
 ## 0.88.0 - 2026-03-27
 
 ### Added